Moved THRFormsChildHealthAggregationHelper to own file
